handler separation init
