+++ server/src/handlers/http/modal/ingest/mod.rs
@@ -1,0 +1,3 @@
+pub mod ingester_logstream;
+pub mod ingester_rbac;
+pub mod ingester_role;
