+++ server/src/handlers/http/modal/mod.rs
@@ -20,6 +20,9 @@ pub mod ingest_server;
 pub mod query_server;
 pub mod server;
 pub mod ssl_acceptor;
+pub mod utils;
+pub mod ingest;
+pub mod query;
 
 use std::sync::Arc;
