+++ server/src/handlers/http/modal/server.rs
@@ -60,6 +60,7 @@ use crate::{
 
 // use super::generate;
 use super::generate;
+use super::query::querier_role;
 use super::ssl_acceptor::get_ssl_acceptor;
 use super::OpenIdClient;
 use super::ParseableServer;
@@ -384,25 +385,6 @@ impl Server {
                                     .authorize_for_stream(Action::GetCacheEnabled),
                             ),
                     )
-                    .service(
-                        web::resource("/hottier")
-                            // PUT "/logstream/{logstream}/hottier" ==> Set hottier for given logstream
-                            .route(
-                                web::put()
-                                    .to(logstream::put_stream_hot_tier)
-                                    .authorize_for_stream(Action::PutHotTierEnabled),
-                            )
-                            .route(
-                                web::get()
-                                    .to(logstream::get_stream_hot_tier)
-                                    .authorize_for_stream(Action::GetHotTierEnabled),
-                            )
-                            .route(
-                                web::delete()
-                                    .to(logstream::delete_stream_hot_tier)
-                                    .authorize_for_stream(Action::DeleteHotTierEnabled),
-                            ),
-                    ),
             )
     }
 
@@ -456,14 +438,14 @@ impl Server {
             .service(
                 // PUT, GET, DELETE Roles
                 resource("/{name}")
-                    .route(web::put().to(role::put).authorize(Action::PutRole))
+                    .route(web::put().to(querier_role::put).authorize(Action::PutRole))
                     .route(web::delete().to(role::delete).authorize(Action::DeleteRole))
                     .route(web::get().to(role::get).authorize(Action::GetRole)),
             )
     }
 
     // get the user webscope
-    pub fn get_user_webscope() -> Scope {
+    fn get_user_webscope() -> Scope {
         web::scope("/user")
             .service(
                 web::resource("")
