+++ server/src/handlers/http/modal/utils/rbac_utils.rs
@@ -1,0 +1,18 @@
+use crate::{option::CONFIG, storage::{self, ObjectStorageError, StorageMetadata}};
+
+
+pub async fn get_metadata() -> Result<crate::storage::StorageMetadata, ObjectStorageError> {
+    let metadata = CONFIG
+        .storage()
+        .get_object_store()
+        .get_metadata()
+        .await?
+        .expect("metadata is initialized");
+    Ok(metadata)
+}
+
+pub async fn put_metadata(metadata: &StorageMetadata) -> Result<(), ObjectStorageError> {
+    storage::put_remote_metadata(metadata).await?;
+    storage::put_staging_metadata(metadata)?;
+    Ok(())
+}
