+++ server/src/handlers/http/rbac.rs
@@ -24,20 +24,20 @@ use crate::{
     rbac::{
         map::roles,
         role::model::DefaultPrivilege,
-        user::{self, User as ParseableUser},
+        user::{self},
         Users,
     },
-    storage::{self, ObjectStorageError, StorageMetadata},
+    storage::{self, ObjectStorageError},
     validator::{self, error::UsernameValidationError},
 };
 use actix_web::{http::header::ContentType, web, Responder};
 use http::StatusCode;
 use tokio::sync::Mutex;
 
-use super::cluster::{
+use super::{cluster::{
     sync_password_reset_with_ingestors, sync_user_creation_with_ingestors,
     sync_user_deletion_with_ingestors,
-};
+}, modal::utils::rbac_utils::{get_metadata, put_metadata}};
 
 // async aware lock for updating storage metadata and user map atomicically
 static UPDATE_LOCK: Mutex<()> = Mutex::const_new(());
@@ -76,57 +76,43 @@ pub async fn post_user(
 ) -> Result<impl Responder, RBACError> {
     let username = username.into_inner();
 
-    let mut generated_password = String::default();
     let mut metadata = get_metadata().await?;
-    if CONFIG.parseable.mode == Mode::Ingest {
-        if let Some(body) = body {
-            let user: ParseableUser = serde_json::from_value(body.into_inner())?;
-            let _ = storage::put_staging_metadata(&metadata);
-            let created_role = user.roles.clone();
-            Users.put_user(user.clone());
-            Users.put_role(&username, created_role.clone());
-        }
+
+    validator::user_name(&username)?;
+    let roles: HashSet<String> = if let Some(body) = body {
+        serde_json::from_value(body.into_inner())?
     } else {
-        validator::user_name(&username)?;
-        let roles: HashSet<String> = if let Some(body) = body {
-            serde_json::from_value(body.into_inner())?
-        } else {
-            return Err(RBACError::RoleValidationError);
-        };
+        return Err(RBACError::RoleValidationError);
+    };
 
-        if roles.is_empty() {
-            return Err(RBACError::RoleValidationError);
-        }
-        let _ = UPDATE_LOCK.lock().await;
-        if Users.contains(&username)
-            || metadata
-                .users
-                .iter()
-                .any(|user| user.username() == username)
-        {
-            return Err(RBACError::UserExists);
-        }
-
-        let (user, password) = user::User::new_basic(username.clone());
-
-        generated_password = password;
-        metadata.users.push(user.clone());
-
-        put_metadata(&metadata).await?;
-        let created_role = roles.clone();
-        Users.put_user(user.clone());
-
-        if CONFIG.parseable.mode == Mode::Query {
-            sync_user_creation_with_ingestors(user, &Some(roles)).await?;
-        }
-        put_role(
-            web::Path::<String>::from(username.clone()),
-            web::Json(created_role),
-        )
-        .await?;
+    if roles.is_empty() {
+        return Err(RBACError::RoleValidationError);
+    }
+    let _ = UPDATE_LOCK.lock().await;
+    if Users.contains(&username)
+        || metadata
+            .users
+            .iter()
+            .any(|user| user.username() == username)
+    {
+        return Err(RBACError::UserExists);
     }
 
-    Ok(generated_password)
+    let (user, password) = user::User::new_basic(username.clone());
+
+    metadata.users.push(user.clone());
+
+    put_metadata(&metadata).await?;
+    let created_role = roles.clone();
+    Users.put_user(user.clone());
+
+    put_role(
+        web::Path::<String>::from(username.clone()),
+        web::Json(created_role),
+    )
+    .await?;
+
+    Ok(password)
 }
 
 // Handler for POST /api/v1/user/{username}/generate-new-password
@@ -136,46 +122,26 @@ pub async fn post_gen_password(username: web::Path<String>) -> Result<impl Respo
     let mut new_password = String::default();
     let mut new_hash = String::default();
     let mut metadata = get_metadata().await?;
-    if CONFIG.parseable.mode == Mode::Ingest {
-        let _ = storage::put_staging_metadata(&metadata);
-        if let Some(user) = metadata
-            .users
-            .iter_mut()
-            .filter_map(|user| match user.ty {
-                user::UserType::Native(ref mut user) => Some(user),
-                _ => None,
-            })
-            .find(|user| user.username == username)
-        {
-            new_hash.clone_from(&user.password_hash);
-        } else {
-            return Err(RBACError::UserDoesNotExist);
-        }
-        Users.change_password_hash(&username, &new_hash);
+
+    let _ = UPDATE_LOCK.lock().await;
+    let user::PassCode { password, hash } = user::Basic::gen_new_password();
+    new_password.clone_from(&password);
+    new_hash.clone_from(&hash);
+    if let Some(user) = metadata
+        .users
+        .iter_mut()
+        .filter_map(|user| match user.ty {
+            user::UserType::Native(ref mut user) => Some(user),
+            _ => None,
+        })
+        .find(|user| user.username == username)
+    {
+        user.password_hash.clone_from(&hash);
     } else {
-        let _ = UPDATE_LOCK.lock().await;
-        let user::PassCode { password, hash } = user::Basic::gen_new_password();
-        new_password.clone_from(&password);
-        new_hash.clone_from(&hash);
-        if let Some(user) = metadata
-            .users
-            .iter_mut()
-            .filter_map(|user| match user.ty {
-                user::UserType::Native(ref mut user) => Some(user),
-                _ => None,
-            })
-            .find(|user| user.username == username)
-        {
-            user.password_hash.clone_from(&hash);
-        } else {
-            return Err(RBACError::UserDoesNotExist);
-        }
-        put_metadata(&metadata).await?;
-        Users.change_password_hash(&username, &new_hash);
-        if CONFIG.parseable.mode == Mode::Query {
-            sync_password_reset_with_ingestors(&username).await?;
-        }
+        return Err(RBACError::UserDoesNotExist);
     }
+    put_metadata(&metadata).await?;
+    Users.change_password_hash(&username, &new_hash);
 
     Ok(new_password)
 }
@@ -211,14 +177,7 @@ pub async fn delete_user(username: web::Path<String>) -> Result<impl Responder, 
     let mut metadata = get_metadata().await?;
     metadata.users.retain(|user| user.username() != username);
 
-    if CONFIG.parseable.mode == Mode::Ingest {
-        let _ = storage::put_staging_metadata(&metadata);
-    } else {
-        put_metadata(&metadata).await?;
-        if CONFIG.parseable.mode == Mode::Query {
-            sync_user_deletion_with_ingestors(&username).await?;
-        }
-    }
+    put_metadata(&metadata).await?;
 
     // update in mem table
     Users.delete_user(&username);
@@ -250,36 +209,12 @@ pub async fn put_role(
         return Err(RBACError::UserDoesNotExist);
     }
 
-    if CONFIG.parseable.mode == Mode::Ingest {
-        let _ = storage::put_staging_metadata(&metadata);
-        // update in mem table
-        Users.put_role(&username.clone(), role.clone());
-    } else {
-        put_metadata(&metadata).await?;
-        // update in mem table
-        Users.put_role(&username.clone(), role.clone());
-        if CONFIG.parseable.mode == Mode::Query {
-            sync_users_with_roles_with_ingestors(&username, &role).await?;
-        }
-    }
+    
+    put_metadata(&metadata).await?;
+    // update in mem table
+    Users.put_role(&username.clone(), role.clone());
 
     Ok(format!("Roles updated successfully for {username}"))
-}
-
-async fn get_metadata() -> Result<crate::storage::StorageMetadata, ObjectStorageError> {
-    let metadata = CONFIG
-        .storage()
-        .get_object_store()
-        .get_metadata()
-        .await?
-        .expect("metadata is initialized");
-    Ok(metadata)
-}
-
-async fn put_metadata(metadata: &StorageMetadata) -> Result<(), ObjectStorageError> {
-    storage::put_remote_metadata(metadata).await?;
-    storage::put_staging_metadata(metadata)?;
-    Ok(())
 }
 
 #[derive(Debug, thiserror::Error)]
